Modularize workers and renderers (#36)
This refactors the worker logic from `hekla-webpack-plugin` and moves it into the core, as a new `WorkQueue` class. Running analysis now starts a batch of workers that will asynchronously consume from a queue of modules. The rendering logic has been refactored into a new `StatusUpdatePlugin`, which can be used from within either the CLI or the webpack plugin.