Add integration tests for hekla-cli package (#42)
This adds a set of basic integration tests for the `hekla-cli` package. The tests work by creating a `dist` folder and running the `analyze` command manually. This means we are skipping the actual bash command line interface, but we are able to test on the layer just below it. The tests assert that the analysis file is created and some basic plugin functionality is executed.

We can put heavier integration test coverage on `hekla-core` to test that specifics of plugin behavior are working as expected, but simple tests like these will probably be good enough for the CLI and webpack plugin, since these call into the `Analyzer` class in `hekla-core`. From the CLI's perspecitive, we just need to know that those methods are being called.